split scriptrunner.
